Tweaked views to fail more gracefully, if any data makes it through without a theme set beforehand
In event_search_filter, using 'blank?' instead of 'empty?' means that if @events is nil, the page still continues rendering. If we use empty, Rails expects an array, and errors out on us.

In calendar view the '.present?' method checks if an event has a theme, before trying to render that theme in snake-case. This stops another rendering error.

Feel free to get rid of these two changes if you're confident that the data will definitely be clean.