Moved processing for ajax map view to server side.
Worthy of note:

Google maps tries to guess the size an infobubble needs to be by looking at all the styles that apply directly to elements inside. This means that any styles defined further up the tree, where you might expect to let the cascade to do work for you won't be taken into account.  This should explain some duplicated styles on the map elements.